+++ src/assets/slides.js
@@ -1,0 +1,26 @@
+import coffee from "../../src/assets/coffee.jpg";
+import coffee_mobile from "../../src/assets/coffee_mobile.jpg";
+import community from "../../src/assets/community.jpg";
+import community_mobile from "../../src/assets/community_mobile.jpg";
+import donut from "../../src/assets/donut.jpg";
+import donut_mobile from "../../src/assets/donut_mobile.jpg";
+
+const slides = [
+    {
+      image: coffee,
+      mobileImage: coffee_mobile,
+      text: 'Coffee is more than just a drink — it’s an art form. As a 3rd Wave café, each bean is ethically sourced and each cup is crafted with care for an elevated experience.'
+    },
+    {
+        image: donut,
+        mobileImage: donut_mobile,
+        text: 'Our deliciously fluffy,  100% plant-based doughnuts are made fresh daily from scratch, with love and using only natural, organic ingredients.'
+    },
+    {
+        image: community,
+        mobileImage: community_mobile,
+      text: 'From eco-friendly practices to local partnerships, we’re committed to creating a sustainable future and building a connected community.'
+    }
+  ];
+
+  export default slides
